chore(tests): tested 'parseSourceCode' and 'parseGeneratedCode' functions directly
