+++ src/utils/parseGeneratedCode.spec.ts
@@ -1,34 +1,21 @@
-import SASjs from './index'
-
-const adapter = new SASjs()
-
-it('should parse SAS9 source code', async (done) => {
-  expect(sampleResponse).toBeTruthy()
-  const parsedSourceCode = (adapter as any).parseSAS9SourceCode(sampleResponse)
-  expect(parsedSourceCode).toBeTruthy()
-  const sourceCodeLines = parsedSourceCode.split('\r\n')
-  expect(sourceCodeLines.length).toEqual(5)
-  expect(sourceCodeLines[0].startsWith('6')).toBeTruthy()
-  expect(sourceCodeLines[1].startsWith('7')).toBeTruthy()
-  expect(sourceCodeLines[2].startsWith('8')).toBeTruthy()
-  expect(sourceCodeLines[3].startsWith('9')).toBeTruthy()
-  expect(sourceCodeLines[4].startsWith('10')).toBeTruthy()
-  done()
-})
+import { parseGeneratedCode } from './index'
 
 it('should parse generated code', async (done) => {
   expect(sampleResponse).toBeTruthy()
-  const parsedGeneratedCode = (adapter as any).parseGeneratedCode(
-    sampleResponse
-  )
+
+  const parsedGeneratedCode = parseGeneratedCode(sampleResponse)
+
   expect(parsedGeneratedCode).toBeTruthy()
+
   const generatedCodeLines = parsedGeneratedCode.split('\r\n')
+
   expect(generatedCodeLines.length).toEqual(5)
   expect(generatedCodeLines[0].startsWith('MPRINT(MM_WEBIN)')).toBeTruthy()
   expect(generatedCodeLines[1].startsWith('MPRINT(MM_WEBLEFT)')).toBeTruthy()
   expect(generatedCodeLines[2].startsWith('MPRINT(MM_WEBOUT)')).toBeTruthy()
   expect(generatedCodeLines[3].startsWith('MPRINT(MM_WEBRIGHT)')).toBeTruthy()
   expect(generatedCodeLines[4].startsWith('MPRINT(MM_WEBOUT)')).toBeTruthy()
+
   done()
 })
